Create SVG module
