+++ lib/anime.esm.js
@@ -75,6 +75,10 @@ const transformsExecRgx = /(\w+)\(([^)]*)\)/g;
 const relativeValuesExecRgx = /^(\*=|\+=|-=)/;
 const whiteSpaceTestRgx = /\s/g;
 
+// Math variables
+
+const pi = Math.PI;
+
 // Strings
 
 function selectString(str) {
@@ -527,8 +531,6 @@ function decomposeValue(val, unit) {
   }
 }
 
-// SVG
-
 // getTotalLength() equivalent for circle, rect, polyline, polygon and line shapes
 // adapted from https://gist.github.com/SebLambla/3e0550c496c236709744
 
@@ -537,7 +539,7 @@ function getDistance(p1, p2) {
 }
 
 function getCircleLength(el) {
-  return Math.PI * 2 * el.getAttribute('r');
+  return pi * 2 * el.getAttribute('r');
 }
 
 function getRectLength(el) {
@@ -553,6 +555,7 @@ function getLineLength(el) {
 
 function getPolylineLength(el) {
   const points = el.points;
+  if (is.und(points)) return;
   let totalLength = 0;
   let previousPos;
   for (let i = 0 ; i < points.numberOfItems; i++) {
@@ -565,10 +568,9 @@ function getPolylineLength(el) {
 
 function getPolygonLength(el) {
   const points = el.points;
+  if (is.und(points)) return;
   return getPolylineLength(el) + getDistance(points.getItem(points.numberOfItems - 1), points.getItem(0));
 }
-
-// Path animation
 
 function getTotalLength(el) {
   if (el.getTotalLength) return el.getTotalLength();
@@ -587,16 +589,17 @@ function setDashoffset(el) {
   return pathLength;
 }
 
-// Motion path
-
 function getParentSvgEl(el) {
   let parentEl = el.parentNode;
   while (is.svg(parentEl)) {
-    if (!is.svg(parentEl.parentNode)) break;
-    parentEl = parentEl.parentNode;
+    const parentNode = parentEl.parentNode;
+    if (!is.svg(parentNode)) break;
+    parentEl = parentNode;
   }
   return parentEl;
 }
+
+// SVG Responsive utils
 
 function getParentSvg(pathEl, svgData) {
   const svg = svgData || {};
@@ -618,6 +621,8 @@ function getParentSvg(pathEl, svgData) {
   }
 }
 
+// Path animation
+
 function getPath(path, percent) {
   const pathEl = is.str(path) ? selectString(path)[0] : path;
   const p = percent || 100;
@@ -631,21 +636,23 @@ function getPath(path, percent) {
   }
 }
 
-function getPathProgress(path, progress, isPathTargetInsideSVG) {
-  function point(offset = 0) {
-    const l = progress + offset >= 1 ? progress + offset : 0;
-    return path.el.getPointAtLength(l);
-  }
-  const svg = getParentSvg(path.el, path.svg);
-  const p = point();
-  const p0 = point(-1);
-  const p1 = point(+1);
-  const scaleX = isPathTargetInsideSVG ? 1 : svg.w / svg.vW;
-  const scaleY = isPathTargetInsideSVG ? 1 : svg.h / svg.vH;
-  switch (path.property) {
-    case 'x': return (p.x - svg.x) * scaleX;
-    case 'y': return (p.y - svg.y) * scaleY;
-    case 'angle': return Math.atan2(p1.y - p0.y, p1.x - p0.x) * 180 / Math.PI;
+function getPathPoint(pathEl, progress, offset = 0) {
+  const length = progress + offset >= 1 ? progress + offset : 0;
+  return pathEl.getPointAtLength(length);
+}
+
+function getPathProgress(pathObject, progress, isPathTargetInsideSVG) {
+  const el = pathObject.el;
+  const parentSvg = getParentSvg(pathObject.el, pathObject.svg);
+  const p = getPathPoint(el, progress, 0);
+  const p0 = getPathPoint(el, progress, -1);
+  const p1 = getPathPoint(el, progress, +1);
+  const scaleX = isPathTargetInsideSVG ? 1 : parentSvg.w / parentSvg.vW;
+  const scaleY = isPathTargetInsideSVG ? 1 : parentSvg.h / parentSvg.vH;
+  switch (pathObject.property) {
+    case 'x': return (p.x - parentSvg.x) * scaleX;
+    case 'y': return (p.y - parentSvg.y) * scaleY;
+    case 'angle': return Math.atan2(p1.y - p0.y, p1.x - p0.x) * 180 / pi;
   }
 }
