+++ tests/visual/media-values.html
@@ -1,18 +1,10 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <title>Test | Media values | anime.js</title>
+  <title>Media values tests | anime.js</title>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-  <meta property="og:title" content="anime.js">
-  <meta property="og:url" content="https://animejs.com">
-  <meta property="og:description" content="Javascript Animation Engine">
-  <meta property="og:image" content="https://animejs.com/documentation/assets/img/icons/og.png">
-  <meta name="twitter:card" content="summary">
-  <meta name="twitter:title" content="anime.js">
-  <meta name="twitter:site" content="@juliangarnier">
-  <meta name="twitter:description" content="Javascript Animation Engine">
-  <meta name="twitter:image" content="https://animejs.com/documentation/assets/img/icons/twitter.png">
+  <link href="assets/css/styles.css" rel="stylesheet">
 </head>
 <body>
   <!-- © https://github.com/anars/blank-audio -->
