+++ tests/visual/svg-lines.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>SVG line tests | anime.js</title>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+  <link href="assets/css/styles.css" rel="stylesheet">
+  <style>
+    body {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      position: absolute;
+      overflow: hidden;
+      width: 100%;
+      height: 100%;
+    }
+    .anim {
+      position: relative;
+      width: 50%;
+    }
+    .anim svg {
+      position: relative;
+      width: 100%;
+      border: 1px solid red;
+    }
+    .anim path {
+      stroke-width: 2;
+    }
+    .item {
+      position: absolute;
+      left: 0;
+      top: 0;
+      width: 30px;
+      height: 30px;
+      margin-left: -15px;
+      margin-top: -15px;
+      background-color: white;
+    }
+  </style>
+</head>
+<body>
+  <div class="anim">
+    <div class="item"></div>
+    <svg xmlns="http://www.w3.org/2000/svg" preserveAspectRatio="xMidYMid slice" viewBox="0 0 600 400">
+      <g fill="none" fill-rule="evenodd" stroke-width="2">
+        <line id="line1" x1="51.5" x2="149.5" y1="51.5" y2="149.5" stroke="#F96F82" stroke-linecap="square"/>
+        <line id="line2" x1="149.5" x2="51.5" y1="51.5" y2="149.5" stroke="#F96F82" stroke-linecap="square"/>
+        <circle id="circle" cx="300" cy="100" r="50" stroke="#FED28B"/>
+        <polygon id="polygon" stroke="#D1FA9E" points="500 130.381 464.772 149 471.5 109.563 443 81.634 482.386 75.881 500 40 517.614 75.881 557 81.634 528.5 109.563 535.228 149"/>
+        <polyline id="polyline" stroke="#7BE6D6" points="63.053 345 43 283.815 95.5 246 148 283.815 127.947 345 63.5 345"/>
+        <path id="path" stroke="#4E7EFC" d="M250 300c0-27.614 22.386-50 50-50s50 22.386 50 50v50h-50c-27.614 0-50-22.386-50-50z"/>
+        <rect id="rect" width="100" height="100" x="451" y="251" stroke="#C987FE" rx="25"/>
+      </g>
+    </svg>
+  </div>
+  <script type="module">
+    import anime from '../../src/anime.js';
+    const dashOffsetAnimation = anime({
+      targets: ['line', 'circle', 'polygon', 'polyline', 'path', 'rect'],
+      strokeDashoffset: [anime.setDashoffset, 0],
+      easing: 'easeInOutSine',
+      duration: 1500,
+      direction: 'alternate',
+      loop: true
+    });
+    const pathEl = anime.path('#path');
+    const pathAnimation = anime({
+      targets: '.item',
+      translateX: pathEl('x'),
+      translateY: pathEl('y'),
+      rotate: pathEl('angle'),
+      easing: 'easeInOutSine',
+      duration: 1500,
+      direction: 'alternate',
+      loop: true
+    });
+  </script>
+</body>
+</html>
